perf(core): split histograms panel into a lazily-loaded chunk
Extract the recharts-backed Plots component into HistogramsPanel.tsx; histograms.tsx is now registration + React.lazy. Moves recharts off the entry chunk. Depends on the Panel Suspense boundary.
